enhance: 拆分AmapController
